Move postgres adapter into packages directory
