+++ packages/api-pls-express-server/app.js
@@ -8,7 +8,7 @@ const compress = require('compression');
 const bodyParser = require('body-parser');
 const log = require('./log');
 const loadResourceModels = require('../../lib/resource-model/load-from-disk');
-const postgresAdapter = require('../../adapters/postgres');
+const postgresAdapter = require('../api-pls-postgres-adapter');
 const ApiRouter = require('../../packages/api-pls-express-router');
 
 const NODE_ENV = process.env.NODE_ENV || 'development';
