+++ packages/api-pls-postgres-adapter/middleware/delete.js
